+++ addOns/client/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Added support for Client Script Authentication when installed in conjunction with the Authentication Helper add-on. 
 
 ## [0.12.0] - 2025-01-24
 ### Fixed